pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = ["we1lman <andrey.velma2005@gmail.com>"]
description = ""
name = "inteam-inc"
package-mode = false
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
aiosqlite = "^0.20.0"
bcrypt = "^4.1.2"
fastapi = "^0.110.1"
fastapi-users = {extras = ["sqlalchemy"], version = "^13.0.0"}
pylint = "^3.1.0"
python = "^3.11"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
uvicorn = "^0.29.0"
pydantic = "^2.7.0"
pydantic-settings = "^2.2.1"
passlib = "^1.7.4"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.group.dev.dependencies]
black = "^24.4.0"
flake8 = "^7.0.0"
pylint = "^3.1.0"
toml-sort = "^0.23.1"
